requirements.txt
requests
python-telegram-bot==20.8
beautifulsoup4
playwright
pytz
Pillow
